fix(dataset): keep `class_id` integral for VOC background images (#2463)
detections_from_xml_obj built class_id with np.array(...) over a list of
indices. For an annotation file with no object elements that list is empty,
so NumPy inferred float64 and DetectionDataset validation rejected the
resulting Detections, making any Pascal VOC dataset that contains an
unannotated image impossible to load.

+++ src/supervision/dataset/formats/pascal_voc.py
@@ -347,7 +347,8 @@ def detections_from_xml_obj(
         if k not in extended_classes:
             extended_classes.append(k)
     class_id = np.array(
-        [extended_classes.index(class_name) for class_name in class_names]
+        [extended_classes.index(class_name) for class_name in class_names],
+        dtype=int,
     )
 
     annotation = Detections(
